Fixed security RLS helper functions
