+++ doc/docs/guides/supabse-indexes.md
@@ -1,0 +1,11 @@
+---
+id: supabase-indexes
+title: Database Indexes
+---
+
+These indexes are added to improve the performance of foreign keys in database tables:
+
+```sql
+CREATE INDEX ON "chats"."messages" USING btree ("authorId");
+CREATE INDEX ON "chats"."messages" USING btree ("roomId");
+```
