--- example/lib/constants.dart
@@ -1,1 +1,0 @@
-const String storageBucket = 'messages_files';
