+++ example/lib/main.dart
@@ -2,9 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_supabase_chat_core/flutter_supabase_chat_core.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
-import 'color_schemes.g.dart';
-import 'home.dart';
-import 'rooms.dart';
+import 'src/pages/home.dart';
+import 'src/theme/color_schemes.dart';
 import 'supabase_options.dart';
 
 void main() async {
