+++ example/lib/src/pages/auth.dart
