+++ example/lib/src/pages/home.dart
@@ -1,15 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
-import 'package:flutter_supabase_chat_core/flutter_supabase_chat_core.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
-import 'package:timeago/timeago.dart' as timeago;
 
 import 'auth.dart';
-import 'chat.dart';
 import 'rooms.dart';
 import 'users.dart';
-import 'util.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -48,70 +43,6 @@ class _HomePageState extends State<HomePage> {
 
   void logout() async {
     await Supabase.instance.client.auth.signOut();
-  }
-
-  Widget _buildAvatar(types.Room room) {
-    var color = Colors.transparent;
-    types.User? otherUser;
-
-    if (room.type == types.RoomType.direct) {
-      try {
-        otherUser = room.users.firstWhere(
-          (u) => u.id != _user!.id,
-        );
-
-        color = getUserAvatarNameColor(otherUser);
-      } catch (e) {
-        // Do nothing if the other user is not found.
-      }
-    }
-
-    final hasImage = room.imageUrl != null;
-    final name = room.name ?? '';
-    final Widget child = CircleAvatar(
-      backgroundColor: hasImage ? Colors.transparent : color,
-      backgroundImage: hasImage ? NetworkImage(room.imageUrl!) : null,
-      radius: 20,
-      child: !hasImage
-          ? Text(
-              name.isEmpty ? '' : name[0].toUpperCase(),
-              style: const TextStyle(color: Colors.white),
-            )
-          : null,
-    );
-    if (otherUser == null) {
-      return Container(
-        margin: const EdgeInsets.only(right: 16),
-        child: child,
-      );
-    }
-
-    return Container(
-      margin: const EdgeInsets.only(right: 16),
-      child: UserOnlineStatusWidget(
-        uid: otherUser.id,
-        builder: (status) => Stack(
-          alignment: Alignment.bottomRight,
-          children: [
-            child,
-            if (status == UserOnlineStatus.online)
-              Container(
-                width: 10,
-                height: 10,
-                margin: const EdgeInsets.only(right: 3, bottom: 3),
-                decoration: BoxDecoration(
-                  color: Colors.green,
-                  shape: BoxShape.circle,
-                  border: Border.all(
-                    color: Colors.white,
-                    width: 2,
-                  ),
-                ),
-              ),
-          ],
-        ),
-      ),
-    );
   }
 
   @override
