+++ example/lib/src/theme/color_schemes.dart
