+++ example/utils/sql/03_database_policy.sql
@@ -17,7 +17,7 @@ CREATE OR REPLACE FUNCTION chats.is_auth()
     RETURNS boolean
     LANGUAGE 'plpgsql'
     COST 100
-    VOLATILE NOT LEAKPROOF SECURITY DEFINER
+    VOLATILE NOT LEAKPROOF SECURITY INVOKER
     SET search_path = ''
 AS $BODY$
 BEGIN
@@ -29,7 +29,7 @@ CREATE OR REPLACE FUNCTION chats.is_owner(user_id uuid)
     RETURNS boolean
     LANGUAGE 'plpgsql'
     COST 100
-    VOLATILE NOT LEAKPROOF SECURITY DEFINER
+    VOLATILE NOT LEAKPROOF SECURITY INVOKER
     SET search_path = ''
 AS $BODY$
 BEGIN
@@ -41,7 +41,7 @@ CREATE OR REPLACE FUNCTION chats.is_member(members uuid[])
     RETURNS boolean
     LANGUAGE 'plpgsql'
     COST 100
-    VOLATILE NOT LEAKPROOF SECURITY DEFINER
+    VOLATILE NOT LEAKPROOF SECURITY INVOKER
     SET search_path = ''
 AS $BODY$
 BEGIN
@@ -53,7 +53,7 @@ CREATE OR REPLACE FUNCTION chats.is_chat_member(room_id bigint)
     RETURNS boolean
     LANGUAGE 'plpgsql'
     COST 100
-    VOLATILE NOT LEAKPROOF SECURITY DEFINER
+    VOLATILE NOT LEAKPROOF SECURITY INVOKER
     SET search_path = ''
 AS $BODY$
 DECLARE
